requirements.txt
streamlit
pandas
folium
streamlit-folium
plotly
Pillow